\name{logit}
\alias{logit}
\title{logit transformation of the vector}
\description{
 logit the vector: log(x)-log(1-x)
}
\usage{
logit(x)
}
\arguments{
  \item{x}{vector to be converted to logit scale}
}

\value{
vector of logit transformed data
}
\author{ Vasyl vasyl@unc.edu }
\keyword{ utilities }

