\name{logiti}
\alias{logiti}
\title{invert logit transformation of the vector}
\description{
 logit the vector: 1/(1+exp(-x))
}
\usage{
logiti(x)
}
\arguments{
  \item{x}{vector to be converted back from logit scale}
}

\value{
vector of inverted logit transformed data
}
\author{ Vasyl vasyl@unc.edu }
\keyword{ utilities }

